Implement save to spotify buttom
